--- a/.github/workflows/test.yml
+++ b/.github/workflows/test.yml
@@ -28,7 +28,7 @@ jobs:
       - name: Setup Node.js
         uses: actions/setup-node@v4
         with:
-          node-version: 16.14.0
+          node-version: 20.15.1
 
       - name: Setup pnpm
         uses: pnpm/action-setup@v4