--- a/packages/twig-new/docker-compose.yml
+++ b/packages/twig-new/docker-compose.yml
@@ -6,4 +6,4 @@ services:
             - 8082:80
         volumes:
             - ./dist/components:/opt/drupal/modules/ilo_base_theme_companion/dist/components/
-            - ./dist/global/styles.css:/opt/drupal/modules/ilo_base_theme_companion/dist/index.css
+            - ./dist/styles/index.css:/opt/drupal/modules/ilo_base_theme_companion/dist/index.css