--- a/packages/twig-new/package.json
+++ b/packages/twig-new/package.json
@@ -3,7 +3,11 @@
   "private": true,
   "version": "0.0.0",
   "type": "module",
+  "files": [
+    "dist/**/*"
+  ],
   "scripts": {
+    "build:lib": "vite build",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build"
   },
@@ -20,6 +24,8 @@
     "@storybook/test": "^8.1.9",
     "drupal-attribute": "^1.0.2",
     "drupal-twig-extensions": "^1.0.0-beta.5",
+    "glob": "^10.3.10",
+    "rollup-plugin-copy": "^3.4.0",
     "storybook": "^8.1.9",
     "twig": "^1.16.0",
     "vite": "^5.2.0",
@@ -31,6 +37,7 @@
     "@ilo-org/icons": "workspace:*",
     "@ilo-org/maestro": "workspace:*",
     "@ilo-org/styles": "workspace:*",
+    "@ilo-org/themes": "workspace:*",
     "@ilo-org/utils": "workspace:*",
     "@popperjs/core": "^2.11.8",
     "video.js": "^7.21.2",