--- a/packages/twig-new/package.json
+++ b/packages/twig-new/package.json
@@ -36,7 +36,6 @@
     "@ilo-org/fonts": "workspace:*",
     "@ilo-org/icons": "workspace:*",
     "@ilo-org/maestro": "workspace:*",
-    "@ilo-org/themes": "workspace:*",
     "@ilo-org/styles": "workspace:*",
     "@ilo-org/themes": "workspace:*",
     "@ilo-org/utils": "workspace:*",