Fixed #1115 

This PR implements the build script for the new twig package using vite. It is now around 6x faster than the previous implementation.

Before
```
pnpm build:lib  30.63s user 2.32s system 161% cpu 20.359 total
```

After
```
pnpm build:lib  5.49s user 0.91s system 136% cpu 4.695 
```

Compared to the previous twig package, there are a few notable changes
- The JS is output as ES modules instead of self contained UMD modules - this means that some of the JS files import a shared chunk file with common code. As long as all the JS files are hosted with the same directory structure as exported in the package, this should work fine in the browser.
- The build output follows the project structure - the current implementation deals with sub components in an inconsistent way. For example for the `tag` sub component that is under `tags` in the project, the twig file will be copied to `tags` and then the pattern and css files are copied to `tag`. In the new implementation, the output files will be follow the structure of the project so in this case `tag` will be placed under `tags`.
- Component CSS files are no longer outputted - The CSS file names don't match the names of the components in the twig package so they can't correctly be colocated in the build output. Since both the base theme and ilo.org site use the single combined style file, the twig package only outputs that now.