@bashlk 2024-07-16
@GGKapanadze @justintemps I have an uneasy feeling switching to ES modules. It has great browser support but it is still not 100%. On one hand, shared code is not duplicated. But on the other, the output modules are not fully minified and as far as I can see, vite / rollup can't be made to output minified ES modules.

If we choose not to use ES modules, we have to switch back to using pure rollup. Vite doesn't support outputting several UMD / IIFE modules so we have to go back to feeding several configs to rollup.