@bashlk 2024-07-22
After giving it some thought, I think it is best if we bundle each behavior file separately as we do with the current twig package. That will yield completely standalone JS files for each component that are the most minified and the most compatible across all browsers.

Sadly, this means that we will have to continue using rollup but I think I can improve the build scripts used in the old twig package. Will make the adjustments as soon as I can.