@bashlk 2024-07-22
However, on the other hand, the JS files from the twig package will be processed by Drupal before they are sent to the user (https://www.drupal.org/node/3301716) and it seems Drupal supports ES modules for its JS files. It apparently also minifies the JS files 🤔

So I will test how ES modules work in Drupal. If it works, we can leave this PR as is.