@ademarco 2024-07-23
@bashlk I've tested it and we can use ES modules in Drupal.