@bashlk 2024-07-23
After exploring this a bit further with @ademarco we think that ES modules is not the way to go.

When a behavior JS file is marked as a module on Drupal, Drupal skips the aggregating the JS and instead requires the client to download each behavior file separately which is not performant. This combined with less browser support and the larger size of each behavior file when output from Vite as a ES module, leads it to be a poor choice.

So we proceed with iife. cc @GGKapanadze 

(umd is a combination of several formats which we don't need - https://rollupjs.org/configuration-options/#output-format - so iife is the way to go)