@bashlk 2024-07-25
@justintemps @GGKapanadze I reworked the build scripts to use only rollup, output IIFE modules and to output all the component CSS files to a separate folder called styles. Let me know if it looks alright to you.